pyproject.toml
[tool.poetry]
name = "signpost"
version = "0.0.0"
description = "A self-hosted URL redirect service"
authors = ["Oscar Peace <oscar@oscarcp.net>"]
license = "GPLv3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
flask-limiter = "^3.12"
flask-compress = "^1.17"
qrcode = "^8.2"
pillow = "^11.2.1"
flask-sqlalchemy = "^3.1.1"
flask-htmx = "^0.4.0"
argon2-cffi = "^25.1.0"
flask-wtf = "^1.2.2"
jsonschema = "^4.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
